Add `--skip-if-loaded` for `ecto.load` (#112)
Today, if we want to add the `mix ecto.load`
to a pipeline of tasks, it will fail if the
database structure is already up and it will
not run the next tasks.

For example, let's say I want to build a
pipeline called `ecto.setup` like this:

```
defp aliases do
    [
      "ecto.setup": [
        "ecto.create",
        "ecto.load",
        "ecto.migrate",
        "run priv/repo/seeds.exs"
    ],
```
If the database is already up, the `ecto.create`
doesn't fail and proceed to `ecto.load`. Then, in
`ecto.load` it fails if there tables there and doesn't
execute the next tasks in the pipeline. But not always
having tables in the database is a problem, in this
pipeline would be cool just skip the `ecto.load` if
the tables are already there.

We discussed this option here in the mailing list:
https://groups.google.com/forum/#!topic/elixir-ecto/8aw2DL_c9Yc

We discussed that a good approach would be, before attempting to
load the bump, would be check if there's any migration run. If yes,
it's not a new database, it's loaded. If there are no migrations
 there, we can try to load the dump.

With the `--skip-if-loaded`, when the database is not new, we
can skip the load and succeed in the task. What do you folks think?