Restart connections instead of restarting whole apps
Whenever a migration adds an extension, we have to restart
connections so they rebootstrap. We were used to achieve
this by restarting Postgrex but that would restart the
connections which in turn triggers a race condition as seen
on #119.

This commit takes a different approach by simply restarting
the whole pool whenever a migration on an already connected
repo is running.

Closes #119.