Add logic to support new column type of :identity
This may be used for Postgres version >= 10, and is considered
preferable to using serial columns because of reduced permissions
maintenance.

By default, all :identity column will be bigints. You may provide
optional parameters for `:start_value` and `:increment` to customize the
created sequence.

To override the `GENERATED BY DEFAULT`, use a straight text or long
atom in the column definition:

    add :id, :"bigint GENERATED ALWAYS AS IDENTITY  (START WITH 100 INCREMENT BY 20)"

Configuration:

* Per table: `add :id, :identity, primary_key: true`
* Per repo: `config :app, App.Repo, migration_primary_key: [type: :identity]`

I could not get

    config :app, App.Repo, migration_primary_key: [type: :"int GENERATED BY DEFAULT AS IDENTITY"]

to work, since every table that `references() the primary key id column
when be of type "int GENERATED BY DEFAULT AS IDENTITY" and not "int".

To work around this, you may set your Application config so that
ecto_sql: [postgres_identity_type: :int]

Special thanks to @seivan on Github for the initial PR that got this
started.