Fix bug in Tds adapter with missing () for multi-line fragments
There was a bug with the parens_for_select/1 function where the
matching for SELECT at the start of an expression wasn't being
properly matched if there was leading space.  This resulted
the brackets () missing for multi-line fragments.

The fix was to use a regex to match for select at the start of
the line.