fix: parent_as when used at right side of a combining expression
_Notice: it depends on https://github.com/elixir-ecto/ecto/pull/3627_

This patch fixes a bug in which is impossible to compile and execute a
query using `parent_as` at the right side of a union expression (or
any other type of combination). As an example, consider the following
query:

```
  subquery = from(c in Comments, where: parent_as(:posts).id == c.post_id)
  from(p in Posts, where: exists(union_all(subquery, ^subquery)))
```

This query should fail to compile with the following error:

```
** (Ecto.SubQueryError) the following exception happened when compiling a subquery.
    ** (Ecto.QueryError) `parent_as(:post)` can only be used in subqueries in query:
```

After ecto is fixed and the queries compiles, there is a runtime
problem that occurs when you execute that query. It happens because
the alias can't be found when the query is actually
executed:

```
** (ArgumentError) argument error
  :erlang.element(1, nil)
  (ecto_sql 3.6.1) lib/ecto/adapters/postgres/connection.ex:1215: Ecto.Adapters.Postgres.Connection.quote_qualified_name/3
  (ecto_sql 3.6.1) lib/ecto/adapters/postgres/connection.ex:697: Ecto.Adapters.Postgres.Connection.expr/3
  ...
```

Similarly to what has been done in the ecto PR, we modify the code to
include the alias in all combining expressions related to the subquery.