Make `modify/3` with `:null` option reversible
`modify/3` has always accepted the null option, but it never properly
migrated it when it would be run backwards. Since modifying a column
only requires the `:null` option if it has to be changed, Ecto should be
safe to assume the opposite of the passed value is the reverse of it.