Add support for index directions
As discussed [on the mailing list], this PR add support for specifying the index
direction when creating a new index.

[on the mailing list]:
https://groups.google.com/g/elixir-ecto/c/uYFE2evwsW4/m/2q5QwQH2BAAJ

Most of the changes are pretty straightforward, but in the Tds and Postgres
adapters, I did have to split up the old `index_expr/1` into two different
functions: one to compute the index expression and one to compute the include
expression. Previously they shared the same function, but now that the domains
of the functions are different I don't think they should.

I also added two more types to `Ecto.Migration` to make sure the type of the
`columns` field is correct.

Finally, I chose to ignore the sorting direction when generating the index name,
so that functionality also had to be slightly amended.