Allow configuring stacktrace depth for logging
A lot of times, more stack items can be helpful in identifying the
source of a query.

This commit demonstrates how we can support it. But the API has
problems:

1. the `:stacktrace` option is used not only by logging but also
for general telemetry events. Making it an integer doesn't make sense for
telemetry (as the full stacktrace is published there anyway).

2. the `:stacktrace` option is overriden Ecto-side, so Ecto.SQL can't
accept integers on a per-query basis, only on global configuration.